[mdl-ripple] Update to mixin pattern and some cleanup.
