+++ packages/mdl-ripple/package.json
@@ -3,5 +3,6 @@
   "version": "1.0.0",
   "license": "MIT",
   "dependencies": {
+    "mdl-base-component": "^1.0.0"
   }
 }
